セカンドデプロイ

+++ app/controllers/public/favorites_controller.rb
@@ -1,7 +1,7 @@
 class Public::FavoritesController < ApplicationController
   def create
     post = Post.find(params[:post_id])
-    favorite = current_user.favorites.new(post_id: post.id)
+    favorite = current_user.favorites.find_or_create_by(post_id: post.id)
     favorite.save
     redirect_to request.referer
   end
@@ -9,7 +9,7 @@ class Public::FavoritesController < ApplicationController
   def destroy
     post = Post.find(params[:post_id])
     favorite = current_user.favorites.find_by(post_id: post.id)
-    favorite.destroy
+    favorite.destroy if favorite
     redirect_to request.referer
   end
 end
